// frontend/src/api/models/UpdateCanvasImageRequest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UpdateCanvasImageRequest = {
    uri?: string;
    'x'?: number | null;
    'y'?: number | null;
    width?: number | null;
    height?: number | null;
    left?: number | null;
    right?: number | null;
    top?: number | null;
    bottom?: number | null;
    size?: number | null;
};